added Order requirement for OrderItem

+++ order-management-service/src/main/java/com/spaghettininjas/yaposs/order/processing/repository/item/OrderItemSpecification.java
@@ -7,11 +7,11 @@ import java.math.BigDecimal;
 public class OrderItemSpecification {
     private OrderItemSpecification() {}
 
-    public static Specification<OrderItem> priceLessThan(BigDecimal priceCeiling) {
+    public static Specification<OrderItem> priceLessThan(Double priceCeiling) {
         return (root, query, builder) -> builder.lessThan(root.get("price"), priceCeiling);
     }
 
-    public static Specification<OrderItem> priceGreaterThan(BigDecimal priceFloor) {
+    public static Specification<OrderItem> priceGreaterThan(Double priceFloor) {
         return (root, query, builder) -> builder.greaterThan(root.get("price"), priceFloor);
     }
 }
